refactor: flatten lib/ structure to standard Flutter template layout
Move files from nested app/features/shared dirs into flat top-level folders. Update all imports. Add converter_screen widget tests.

Made-with: Cursor
