--- lib/features/converter/domain/entities/conversion_result.dart
@@ -1,13 +1,0 @@
-import '../enums/image_format.dart';
-
-class ConversionResult {
-  const ConversionResult({
-    required this.outputPath,
-    required this.format,
-    required this.sizeBytes,
-  });
-
-  final String outputPath;
-  final ImageFormat format;
-  final int sizeBytes;
-}
