--- lib/features/converter/domain/usecases/pick_image_usecase.dart
@@ -1,10 +1,0 @@
-import '../entities/image_asset.dart';
-import '../repositories/converter_repository.dart';
-
-class PickImageUseCase {
-  const PickImageUseCase(this._repository);
-
-  final ConverterRepository _repository;
-
-  Future<ImageAsset?> call() => _repository.pickImage();
-}
