--- lib/features/converter/domain/usecases/save_result_usecase.dart
@@ -1,11 +1,0 @@
-import '../entities/conversion_result.dart';
-import '../repositories/converter_repository.dart';
-
-class SaveResultUseCase {
-  const SaveResultUseCase(this._repository);
-
-  final ConverterRepository _repository;
-
-  Future<String> call(ConversionResult result) =>
-      _repository.saveToGallery(result);
-}
