+++ lib/usecases/pick_image_usecase.dart
@@ -1,0 +1,10 @@
+import '../models/image_asset.dart';
+import '../repositories/converter_repository.dart';
+
+class PickImageUseCase {
+  const PickImageUseCase(this._repository);
+
+  final ConverterRepository _repository;
+
+  Future<ImageAsset?> call() => _repository.pickImage();
+}
