+++ lib/usecases/save_result_usecase.dart
@@ -1,0 +1,11 @@
+import '../models/conversion_result.dart';
+import '../repositories/converter_repository.dart';
+
+class SaveResultUseCase {
+  const SaveResultUseCase(this._repository);
+
+  final ConverterRepository _repository;
+
+  Future<String> call(ConversionResult result) =>
+      _repository.saveToGallery(result);
+}
