+++ lib/utils/result.dart
