+++ lib/widgets/conversion_status_banner.dart
