+++ test/app_test.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:general_converter/app/app.dart';
+import 'package:general_converter/app.dart';
 
 void main() {
   testWidgets('App renders converter screen', (WidgetTester tester) async {
